Migrate to the update app structure
The new app structure is cleaner.

Additionally, this change introduces a new way of reading the config file.
If it doesn’t exist, it just uses the defaults.
If it exists, it set the “config” flag and tries to read from it any flag that might be present.
The reason for using the NewFlagSet function, is that it allows to customize how to behave after an error.
The default Parse function exits (os.Exit(2)) on an error.

Related to #55