Implement gather move operation
It’s essentially a move operation.

Two changes were needed:
- The source folder is no longer static as in Scatter, as there can be multiple source disks
We solve this by adding a workdir property to the command struct to indicate the source
disk path.
For this to work, the getFolders function was modified to populate a Location field in the
Item struct, so when we build the command we use this field.
For Scatter, we use operation.SourceDiskName for all the commands.

- A specific message to signal the gather operation has finished
With Scatter, we use transferFinished, which would work fine, except for the fact that for Gather, we need
to reposition the UI to the main gather page.

Related to #64