Implement Multi-Op resume
This allows the user to refresh the page/restart the browser and have an updated status of any operation
that may be in progress.

The core idea is to have a getStatus api, which returns the current status of the server (what operation
is running)

Based on that response, the UI is redirected towards the appropriate page (route).

Additionally, all the links in the menu are disabled whenever an operation is in progress, to avoid any
side-effects of the user clicking around :)

Related to #64