Implement /proc based rsync handling
We now monitor rsync via /proc, rather than parsing the output
from go/exec

This allows us to recover from any error the rsync command may
generate and decide on an action course based on what the error
is.

But there are some considerations:

It returns code 23 when a permission denied error was previously
thrown.

This means that probably, the file/folder wasn't copied to the
destination, so I shouldn't perform deletions for that command.

We need to do some more thought on this matter.

Ref #125