Implement history file conversion routine
This converts history files from an older version to the current.

It does it steps: v1 -> v2 -> v3 -> so on ...

That's why we have an array of converters, each one knows how to
convert to a specific version, so we just run them one after another,
if needed.

For example, if we're currently a v3 and we find a v1 file, it first
converts v1 to v2 and only then v2 to v3.

Ref #125