Improve handling of rsync error 23
When rsync finds an issue while running, it sends the error to stderr,
but continues execution.

When the command has completed, it ends with an error 23.

This is what we do now. Any stderr message is logged and if the rsync
command ends with code 23, we flag it, DO NOT remove the source folder
(this applies to Gather and Scatter/Move) and continue execution.

The UI makes a distinction for these flagged commands and offers an
option to delete the source folder, once the user had the chance to
check/solve the issue that created the code 23.

Ref #125