Fix planning stage's calculation of space
A recent switch to uint64 for all size/space fields had as a
consequence that calculating differences could be a negative number
which wrapped around uint64's max.

The calculation happened in the knapsack algorithm, when subtracting
the reserved space from the disk's free space.

The solution is to switch back to int64.

Ref #140